% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{launch_runbisonpic}
\alias{launch_runbisonpic}
\title{Run runbisonpic App}
\usage{
launch_runbisonpic()
}
\description{
This function will launch the app locally.
}
\examples{
\dontrun{
launch_runbisonpic()
}
}
