% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runbisonpic-package.R
\docType{package}
\name{runbisonpic-package}
\alias{runbisonpic}
\alias{runbisonpic-package}
\title{runbisonpic: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})

}
\keyword{internal}
